feat: migrate DacPicker.spec
